Extract withStrictImportResolution into a shared helper
S8985's decorator duplicated the exact settings-override block already
used by S8980 (~20 lines: the STRICT_IMPORT_RESOLUTION_SETTINGS constant
and the withStrictImportResolution wrapper), flagged as new-code
duplication. Move it to helpers/testing-library.ts and have both rules
import it instead of keeping their own copy.
